+++ SampleCode/Android/app/src/main/java/com/netease/meetinglib/demo/view/SettingsFragment.java
@@ -6,6 +6,7 @@
 package com.netease.meetinglib.demo.view;
 
 
+import android.util.Log;
 import android.widget.Toast;
 
 import androidx.lifecycle.ViewModelProviders;
@@ -15,6 +16,7 @@ import com.netease.meetinglib.demo.SdkAuthenticator;
 import com.netease.meetinglib.demo.base.BaseFragment;
 import com.netease.meetinglib.demo.databinding.FragmentSettingBinding;
 import com.netease.meetinglib.demo.viewmodel.SettingsViewModel;
+import com.netease.meetinglib.sdk.NECallback;
 import com.netease.meetinglib.sdk.NEMeetingError;
 import com.netease.meetinglib.sdk.control.NEControlParams;
 
@@ -25,8 +27,11 @@ public class SettingsFragment extends BaseFragment<FragmentSettingBinding> {
     @Override
     protected void initView() {
         mViewModel = ViewModelProviders.of(this).get(SettingsViewModel.class);
+        mViewModel.getBeautyFaceValue((resultCode, resultMsg, resultData) -> Log.d("SettingsFragment", "initView:getBeautyFaceValue =  " +resultData ));
+
         binding.btnLogout.setOnClickListener(view -> SdkAuthenticator.getInstance().logout(true));
         binding.btnOpenControl.setOnClickListener(v -> mViewModel.openController(new NEControlParams(SdkAuthenticator.getAccount()), null, this::onOpenControllerCallBack));
+        binding.btnOpenBeauty.setOnClickListener(v -> mViewModel.openBeautyUI((resultCode, resultMsg, resultData) -> Toast.makeText(getActivity(), "进入美颜预览#" + resultMsg, Toast.LENGTH_SHORT).show()));
         binding.btnMeetingSettings.setOnClickListener(v -> MeetingSettingsActivity.start(getActivity()));
     }